add commits

+++ cocos/3d/CCBundleReader.h
@@ -87,7 +87,8 @@ public:
     /**
      * Returns the position of the file pointer.
      */
-    long int tell();
+    //long int tell();
+    ssize_t tell();
 
     /**
      * Sets the position of the file pointer.
